+++ src/extension.ts
@@ -15,7 +15,9 @@
 // You should have received a copy of the GNU General Public License
 // along with cph-ng.  If not, see <https://www.gnu.org/licenses/>.
 
+import { install } from 'source-map-support';
 import ExtensionManager from './modules/extensionManager';
 
+install();
 export const activate = ExtensionManager.activate;
 export const deactivate = ExtensionManager.deactivate;